Remove calls to race
